requirements.txt
# Core FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.11.4

# Redis and Queue Management
redis==5.0.1
psutil==5.9.6

# HTTP and Async
httpx==0.27.2
aiohttp==3.10.5
requests==2.32.3

# Data Processing
pandas==2.2.3
numpy==2.2.5

# Environment and Configuration
python-dotenv==1.1.0

# AI and Analysis
# google-generativeai==0.8.3
google-genai==1.15.0

# Database (Supabase)
supabase==2.15.1
postgrest==1.0.1

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.24.0

# Date and Time
python-dateutil==2.9.0.post0
pytz==2025.2

# Development and Testing
pytest==8.3.5
pytest-asyncio==0.24.0

# System Utilities
schedule==1.2.2
click==8.2.0

# Additional utilities
gunicorn==23.0.0
typing-extensions==4.13.2
